Core/Misc: Moved Position to its own file
